fix issues of Tasker Plugin Pref Editor
