--- app/src/main/res/layout/content_edit.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:background="#ffffffff"
-              android:orientation="vertical"
-              tools:showIn="@layout/activity_edit">
-
-    <com.stardust.scriptdroid.ui.edit.CodeMirrorEditor
-        android:id="@+id/editor"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@null"
-        android:bufferType="spannable"
-        android:cursorVisible="true"
-        android:gravity="top|start"
-        android:text=""
-        android:textCursorDrawable="@null"
-        android:textIsSelectable="true"/>
-
-
-    <com.stardust.scriptdroid.ui.edit.completion.InputMethodEnhanceBar
-        android:id="@+id/input_method_enhance_bar"
-        android:layout_width="match_parent"
-        android:layout_height="35dp"
-        android:background="#e7ebec"/>
-
-</LinearLayout>
